refactor: move packet_buffer to protocol
